Collisions now adjacent (#5)
Colliding with a wall now brings you directly adjacent to the wall, rather than the previous lazy method of simply setting speed to 0.  This prevents issues of "levitating" when you approach the ground at high velocity, as well as cases where you could get too close to the wall and get stuck.